Вносит правки по замечаниям наставника
